% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReport.R
\name{getReport}
\alias{getReport}
\title{Fetches the report from the WDFW website and parses it. If url is NA, the current is used}
\usage{
getReport(requested.report = NA, requested.year = NA, requested.url = NA)
}
\arguments{
\item{report}{String, e.g. "Spring" or "Fall"}

\item{year}{Number. If NA}

\item{url}{String. The url for the page to be scraped}
}
\value{
list
}
\description{
Fetches the report from the WDFW website and parses it. If url is NA, the current is used
}
\examples{
getReport())

}
