% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportList.R
\name{getReportList}
\alias{getReportList}
\title{Fetches a page from the WDFW website and scrapes available report links}
\usage{
getReportList(
 
    url = "https://wdfw.wa.gov/fishing/management/columbia-river/compact/other-information"
)
}
\arguments{
\item{url}{String. The url for the page to be scraped}
}
\value{
dataframe
}
\description{
Fetches a page from the WDFW website and scrapes available report links
}
\examples{
getReportList())

}
