% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeTables.R
\name{getTable_}
\alias{getTable_}
\title{Finds a table in the extracted text from a pdf. Internal use}
\usage{
getTable_(p, title.text = "Table 1.", skip = 3, ncols = 9, minchar = 2)
}
\arguments{
\item{p}{A character vector representation of a pdf document output from pdf_text()}

\item{skip}{A number, the number of lines to skip before assigning the first row of data}

\item{ncols}{The number of columns in the table}

\item{minchar}{The minimum number of spaces needed for a break between columns}
}
\value{
A list
}
\description{
Finds a table in the extracted text from a pdf. Internal use
}
\examples{
getTable_(pdf_text("MyPDF.pdf") \%>\% strsplit(split="\n"),title.text="Table 2.",ncols=13)

}
