% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reports}
\alias{Reports}
\title{Links to Joint Staff Reports}
\format{
\subsection{\code{Reports}}{

A data frame with 4 columns:
\describe{
\item{year}{Year report generated}
\item{report}{Spring or Fall}
\item{title}{Report title}
\item{href}{Report URL}
}
}
}
\source{
\url{https://wdfw.wa.gov/fishing/management/columbia-river/compact/other-information}
}
\usage{
Reports
}
\description{
Data frame of current and past Joint Staff Reports
Report ...
}
\keyword{datasets}
