% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeTables.R
\name{unComma}
\alias{unComma}
\title{Remove commas from a string representation of a number}
\usage{
unComma(valueStrings)
}
\arguments{
\item{string_with_comma}{A string}
}
\value{
A number
}
\description{
Remove commas from a string representation of a number
}
\examples{
unComma("24,000")

}
