% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeTables.R
\name{unPct}
\alias{unPct}
\title{Remove percent sign from a string representation of a number}
\usage{
unPct(string_with_percent)
}
\arguments{
\item{string_with_percent}{A string}
}
\value{
A number
}
\description{
Remove percent sign from a string representation of a number
}
\examples{
unPct("24\%")

}
